W160o: MIS-TYPED FLOAT GLOBALS lever. New detector build/wip/W160GlobalT_scan.py flags functions where retail has lwc1/swc1/mul.s at an offset we emit lw/sw plus invented cvt/mtc1/softfloat glue, then GROUPS them by the DAT_ symbol loaded -- a symbol with many consumers is a mis-typed float global whose single declaration fix moves every consumer. Ranked inventory of 510 globals / 420 flagged functions across 139 files written to build/wip/W160GlobalT_inventory.json. Retyped 21 globals (16 in mdlFile, 3 in mt_sceneFunc, 1 each mt_evtMessage/mt_evtMenu): 21 functions improved, 16 window clears, over-window bytes 18092 -> 12092, nd -9214 (mdlFile -8462). LOST 0, new over-window 0, COMPILE_ERROR 0

+++ src/Event/mt_evtMenu.c
@@ -91,7 +91,7 @@ extern code DAT_006a0270[];
 extern u32 DAT_006a0bf8;
 extern u32 DAT_006a0c00;
 extern char DAT_006a0330[];
-extern u32 DAT_007cadc0;
+extern f32 DAT_007cadc0;
 extern u32 DAT_007cadd0;
 extern u32 DAT_007cae18;
 extern u32 DAT_007cafec;
